# docs/openapi.yaml
openapi: 3.1.0
info:
  title: KIOSK API
  version: 0.0.3
...
paths:
...
  /auth/usercode:
    post:
      summary: Login with usercode. Only for testing purposes
      tags:
        - auth
      parameters:
        - in: query
          name: usercode
          schema:
            type: string
          required: true
          description: the usercode with which an user can login to the system
      responses:
        '200':
          description: Login successfull
...